% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.IRTest.R
\name{anova.IRTest}
\alias{anova.IRTest}
\title{Model comparison}
\usage{
\method{anova}{IRTest}(object, ...)
}
\arguments{
\item{object}{An object of \code{"IRTest"}-class.}

\item{...}{Additional objects of \code{"IRTest"}-class to be compared.}
}
\value{
Model-fit indices and results of likelihood ratio test (LRT).
}
\description{
Model comparison
}
\author{
Seewoo Li \email{cu@yonsei.ac.kr}
}
