% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.IRTest.R
\name{best_model}
\alias{best_model}
\title{Determination of the best model}
\usage{
best_model(..., criterion = "HQ")
}
\arguments{
\item{...}{Candidate models}

\item{criterion}{The criterion to be used.}
}
\value{
The best model and model-fit indices.
}
\description{
Determination of the best model
}
