% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.IRTest.R
\name{coef_se}
\alias{coef_se}
\title{Extract Standard Errors of Model Coefficients}
\usage{
coef_se(object, complete = TRUE)
}
\arguments{
\item{object}{An object for which the extraction of standard errors is meaningful.}

\item{complete}{A logical value indicating if the full standard-error vector should be returned.}
}
\value{
Standard errors extracted from the model (\code{object}).
}
\description{
Standard errors of model coefficients calculated by using Fisher information functions.
}
