% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Title}
\usage{
DataGeneration(
  seed = 1,
  N = 2000,
  nitem = 10,
  prob = 0.5,
  d = 1.7,
  sd_ratio = 1,
  a_l = 0.8,
  a_u = 2.5,
  latent_dist = "normal",
  model = 3,
  c_l = 0,
  c_u = 0.2,
  categ
)
}
\arguments{
\item{categ}{}
}
\description{
Title
}
