% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{Density for two component Normal Mixture Distributions}
\usage{
dist(x, prob = 0.5, m = c(0, 0), s = c(1, 1))
}
\arguments{
\item{x}{vector of quantiles.}

\item{prob}{ratio of two distributions.}

\item{m}{vector of means.}

\item{s}{vector of standard deviations.}
}
\value{
Density
}
\description{
Density for two component Normal Mixture Distributions with means equal to m and standard deviations equal to s.
}
