% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2.R
\name{dist2}
\alias{dist2}
\title{Re-parameterized two-component normal mixture distribution}
\usage{
dist2(x, prob = 0.5, d = 0, sd_ratio = 1, overallmean = 0, overallsd = 1)
}
\arguments{
\item{overallsd}{}
}
\description{
Probability density for the re-parameterized two-component normal mixture distribution.
}
