% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_par.R
\name{distribution_par}
\alias{distribution_par}
\title{Recovering original parameters of two component normal mixture distribution}
\usage{
distribution_par(
  prob = 0.5,
  d = 0,
  sd_ratio = 1,
  overallmean = 0,
  overallsd = 1
)
}
\arguments{
\item{prob}{ratio of two distributions.}

\item{d}{the constant.}

\item{sd_ratio}{ratio of standard deviation.}

\item{overallmean}{overall mean.}

\item{overallsd}{overall standard deviation.}
}
\value{
Parameters of two component normal mixture distribution.
m1 & s1 mean the mean and standard deviation of 1st distribution, m2 & s2  mean the mean and standard deviation of 2nd distribution.
}
\description{
Recovering original parameters of two component normal mixture distribution from re-parameterized parameters.
}
