% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnormal.R
\name{dnormal}
\alias{dnormal}
\title{Density for the normal distribution}
\usage{
dnormal(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}
}
\value{
Density
}
\description{
Density for the normal distribution with mean equal to mean and standard deviation equal to sd.
}
