% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHc.R
\docType{data}
\name{GHc}
\alias{GHc}
\title{Gauss-Hermite constants}
\format{
An object of class \code{numeric} of length 21.
}
\usage{
GHc
}
\description{
a \code{vector} that contains Gauss-Hermite constants
}
\keyword{datasets}
