% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{inform_f_test}
\alias{inform_f_test}
\title{Test information function}
\usage{
inform_f_test(x, test)
}
\arguments{
\item{x}{A vector of \eqn{\theta} value(s).}

\item{test}{An object returned from an estimation function.}
}
\value{
A vector of test information values of the same length as \code{x}.
}
\description{
Test information function
}
\author{
Seewoo Li \email{cu@yonsei.ac.kr}
}
