% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTest.R
\name{IRTest}
\alias{IRTest}
\title{IRTest : Item Response Theory Estimation}
\usage{
IRTest(
  initialitem,
  data,
  range = c(-6, 6),
  q = 121,
  model,
  latent_dist = "Normal",
  max_iter = 200,
  threshold = 1e-04,
  bandwidth = "nrd",
  h = NULL
)
}
\arguments{
\item{initialitem}{initial item from DataGeneration.}

\item{data}{data from DataGeneration.}

\item{range}{range of parameter.}

\item{q}{number of quadrature points}

\item{model}{parameter logistic model of item response theory.
1 means one parameter logistic model, 2 means two parameter logistic model, 3 means three parameter logistic model.}

\item{latent_dist}{Type of latent distribution.
Normal, EHM(Empirical Histogram Method), Mixture(two component mixture normal), KDE(Kernel Density estimation), DC(Davidian Curve)}

\item{max_iter}{Maximum number of iterations}

\item{threshold}{Threshold of convergence.}

\item{bandwidth}{bandwidth of a kernel density estimator.}

\item{h}{the number of hyper-parameters in Davidian Curve.}
}
\value{
item parameter estimator and standard error, fk, Number of iteration, prob, d, sd_ration, quadrature points, diff, Ak, Pk,
theta(ability estimator), logL(deviance), bw(bandwidth), Options
}
\description{
IRTest : Item Response Theory Estimation
}
\details{
fk, Ak, Pk etc
}
