% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTest_Cont.R
\name{IRTest_Cont}
\alias{IRTest_Cont}
\title{Item and ability parameters estimation for continuous response items}
\usage{
IRTest_Cont(
  data,
  range = c(-6, 6),
  q = 121,
  initialitem = NULL,
  ability_method = "EAP",
  latent_dist = "Normal",
  max_iter = 200,
  threshold = 1e-04,
  bandwidth = "SJ-ste",
  h = NULL
)
}
\arguments{
\item{data}{A matrix or data frame of item responses where responses are coded as 0 or 1.
Rows and columns indicate examinees and items, respectively.}

\item{range}{Range of the latent variable to be considered in the quadrature scheme.
The default is from \code{-6} to \code{6}: \code{c(-6, 6)}.}

\item{q}{A numeric value that represents the number of quadrature points. The default value is 121.}

\item{initialitem}{A matrix of initial item parameter values for starting the estimation algorithm. The default value is \code{NULL}.}

\item{ability_method}{The ability parameter estimation method.
The available options are Expected \emph{a posteriori} (\code{EAP}) and Maximum Likelihood Estimates (\code{MLE}).
The default is \code{EAP}.}

\item{latent_dist}{A character string that determines latent distribution estimation method.
Insert \code{"Normal"}, \code{"normal"}, or \code{"N"} for the normality assumption on the latent distribution,
\code{"EHM"} for empirical histogram method (Mislevy, 1984; Mislevy & Bock, 1985),
\code{"2NM"} or \code{"Mixture"} for using two-component Gaussian mixture distribution (Li, 2021; Mislevy, 1984),
\code{"DC"} or \code{"Davidian"} for Davidian-curve method (Woods & Lin, 2009),
\code{"KDE"} for kernel density estimation method (Li, 2022),
and \code{"LLS"} for log-linear smoothing method (Casabianca & Lewis, 2015).
The default value is set to \code{"Normal"} to follow the convention.}

\item{max_iter}{A numeric value that determines the maximum number of iterations in the EM-MML.
The default value is 200.}

\item{threshold}{A numeric value that determines the threshold of EM-MML convergence.
A maximum item parameter change is monitored and compared with the threshold.
The default value is 0.0001.}

\item{bandwidth}{A character value that can be used if \code{latent_dist = "KDE"}.
This argument determines the bandwidth estimation method for \code{"KDE"}.
The default value is \code{"SJ-ste"}. See \code{\link{density}} for available options.}

\item{h}{A natural number less than or equal to 10 if \code{latent_dist = "DC" or "LLS"}.
This argument determines the complexity of the distribution.}
}
\description{
This function estimates IRT item and ability parameters when all items are scored dichotomously.
Based on Bock & Aitkin's (1981) marginal maximum likelihood and EM algorithm (EM-MML), this function provides several latent distribution estimation algorithms which could free the normality assumption on the latent variable.
If the normality assumption is violated, application of these latent distribution estimation methods could reflect non-normal characteristics of the unknown true latent distribution,
and, thus, could provide more accurate parameter estimates (Li, 2021; Woods & Lin, 2009; Woods & Thissen, 2006).
}
