% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTest_Mix.R
\name{IRTest_Mix}
\alias{IRTest_Mix}
\title{Title}
\usage{
IRTest_Mix(
  initialitem_D,
  initialitem_P,
  data_D,
  data_P,
  range = c(-6, 6),
  q = 121,
  model_D,
  model_P = "GPCM",
  latent_dist = "Normal",
  max_iter = 200,
  threshold = 1e-04,
  bandwidth = "nrd",
  h = NULL
)
}
\description{
Title
}
