% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.IRTest.R
\name{logLik.IRTest}
\alias{logLik.IRTest}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{IRTest}(object, ...)
}
\arguments{
\item{object}{A \code{IRTest}-class object from which a log-likelihood value is extracted.}

\item{...}{Other arguments.}
}
\value{
Extracted log-likelihood.
}
\description{
Extract Log-Likelihood
}
