% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_item.R
\name{plot_item}
\alias{plot_item}
\title{Plot of item response function}
\usage{
plot_item(x, item.number, type = NULL)
}
\arguments{
\item{x}{A model fit object from either \code{IRTest_Dich}, \code{IRTest_Poly}, or \code{IRTest_Mix}.}

\item{item.number}{A numeric value indicating the item number.}

\item{type}{Type of an item being either \code{"d"} (dichotomous item) or \code{"p"} (polytomous item); \code{item.number=1, type="d"} indicates the first dichotomous item.
This value will be used only when a model fit of mixed-format data is passed onto the function.}
}
\value{
This function returns a plot of item response function.
}
\description{
This function draws a item response function of an item of the fitted model.
}
\author{
Seewoo Li \email{cu@yonsei.ac.kr}
}
