% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.irtest.R
\name{print.IRTest}
\alias{print.IRTest}
\title{Printing the result}
\usage{
\method{print}{IRTest}(x, ...)
}
\arguments{
\item{x}{An object of \code{"IRTest"}-class obtained from either \code{\link{IRTest_Dich}}, \code{\link{IRTest_Poly}}, or \code{\link{IRTest_Mix}}.}

\item{...}{Additional arguments (currently non-functioning).}
}
\value{
Printed texts on the console recommending the usage of \code{summary} function and the direct access to the details using "$" sign.
}
\description{
This function prints the summarized information.
}
\examples{
\donttest{
data <- DataGeneration(N=1000, nitem_P = 8)$data_P

M1 <- IRTest_Poly(data = data, latent_dist = "KDE")

M1
}
}
\author{
Seewoo Li \email{cu@yonsei.ac.kr}
}
