% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Marginal reliability coefficient of IRT}
\usage{
reliability(x)
}
\arguments{
\item{x}{}
}
\value{
Estimated marginal reliability coefficient
}
\description{
Marginal reliability coefficient of IRT
}
