% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.IRTest.R
\name{summary.IRTest}
\alias{summary.IRTest}
\title{Summary of the results}
\usage{
\method{summary}{IRTest}(object, ...)
}
\arguments{
\item{object}{An object of \code{"IRTest"}-class obtained from either \code{\link{IRTest_Dich}}, \code{\link{IRTest_Poly}}, or \code{\link{IRTest_Mix}}.}

\item{...}{Other argument(s) passed on to summarize the results.}
}
\value{
A plot of estimated latent distribution.
}
\description{
These functions summarize the outputs (e.g., convergence of the estimation algorithm, parameter estimates, AIC, etc.).
}
\examples{
data <- DataGeneration(N=1000,
                       nitem_P = 8,
                       latent_dist = "2NM",
                       d = 1.664,
                       sd_ratio = 2,
                       prob = 0.3)$data_P

M1 <- IRTest_Poly(data = data, latent_dist = "KDE")

summary(M1)

}
