% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Generation of artificial item response data}
\usage{
DataGeneration(
  seed = 1,
  N = 2000,
  nitem_D = NULL,
  nitem_P = NULL,
  prob = 0.5,
  d = 1.7,
  sd_ratio = 1,
  a_l = 0.8,
  a_u = 2.5,
  latent_dist = "normal",
  model_D,
  model_P = "GPCM",
  c_l = 0,
  c_u = 0.2,
  categ
)
}
\arguments{
\item{seed}{A numeric value that is used on random sampling.
The seed number can guarantee the replicability of the result.}

\item{N}{a}

\item{nitem_D}{a}

\item{nitem_P}{a}

\item{prob}{a}

\item{d}{a}

\item{sd_ratio}{a}

\item{a_l}{a}

\item{a_u}{a}

\item{latent_dist}{a}

\item{model_D}{a}

\item{model_P}{a}

\item{c_l}{a}

\item{c_u}{a}

\item{categ}{a}
}
\description{
This function generates artificial item response data in which users can specify the types of items and latent distribution.
}
