% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Data Generation}
\usage{
DataGeneration(
  seed = 1,
  N = 2000,
  nitem = 10,
  prob = 0.5,
  d = 1.7,
  sd_ratio = 1,
  a_l = 0.8,
  a_u = 2.5,
  latent_dist = "normal",
  model = 3,
  c_l = 0,
  c_u = 0.2
)
}
\arguments{
\item{seed}{seed number}

\item{N}{Number of observations to be sampled.}

\item{nitem}{Number of items to be sampled.}

\item{prob}{ratio of two distributions.}

\item{d}{the constant.}

\item{sd_ratio}{ratio of standard deviation.}

\item{a_l}{lower limit of a(item discrimination).}

\item{a_u}{upper limit of a(item discrimination).}

\item{latent_dist}{Type of latent distribution.
(beta, normal, chi)}

\item{model}{parameter logistic model of item response theory.
1 means one parameter logistic model, 2 means two parameter logistic model, 3 means three parameter logistic model.}

\item{c_l}{lower limit of c(guessing parameter).}

\item{c_u}{upper limit of c(guessing parameter).}
}
\value{
Data for computer simulation (item, initialitem, theta, data)
}
\description{
Data Generation for Computer Simulation
}
