% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2.R
\name{dist2}
\alias{dist2}
\title{Density for re-parameterized two component Normal Mixture Distributions}
\usage{
dist2(x, prob = 0.5, d = 0, sd_ratio = 1, overallmean = 0, overallsd = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{prob}{ratio of two distributions.}

\item{d}{the constant.}

\item{sd_ratio}{ratio of standard deviation.}

\item{overallmean}{overall mean.}

\item{overallsd}{overall standard deviation.}
}
\value{
Density
}
\description{
Density for re-parameterizedtwo component Normal Mixture Distributions
}
