% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{inform_f_item}
\alias{inform_f_item}
\title{Item information function}
\usage{
inform_f_item(x, test, item = 1, type = "d")
}
\arguments{
\item{x}{A vector of \eqn{\theta} value(s).}

\item{test}{An object returned from an estimation function.}

\item{item}{A natural number indicating the \eqn{n}th item.}

\item{type}{A character value for a mixed format test which determines the item type:
\code{"d"} and \code{"p"} stand for a dichotomous and polytomous item, respectively.}
}
\value{
A vector of the evaluated item information values.
}
\description{
Item information function
}
\author{
Seewoo Li \email{cu@yonsei.ac.kr}
}
