% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LD.R
\name{plot_LD}
\alias{plot_LD}
\title{Plot of Latent Distribution}
\usage{
plot_LD(model, from = -6, to = 6, add = F)
}
\arguments{
\item{model}{parameter logistic model of item response theory.
1 means one parameter logistic model, 2 means two parameter logistic model, 3 means three parameter logistic model.}

\item{from}{lower boundary of theta}

\item{to}{upper boundary of theta}

\item{add}{logical; if TRUE add to an already existing plot; if NA start a new plot taking the defaults for the limits and log-scaling of the x-axis from the previous plot. Taken as FALSE (with a warning if a different value is supplied) if no graphics device is open.}
}
\value{
plot of latent distribution
}
\description{
Plotting of estimated latent density with IRTest
}
