% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LD.R
\name{plot_LD}
\alias{plot_LD}
\title{Plotting the estimated latent distribution}
\usage{
plot_LD(model, from = -6, to = 6, add = F)
}
\arguments{
\item{model}{An object obtained from}

\item{add}{}
}
\description{
This function draws a plot of the estimated latent distribution (the prior distribution of the latent variable).
}
