% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.IRTest.R
\name{print.IRTest_summary}
\alias{print.IRTest_summary}
\title{Printing the summary}
\usage{
\method{print}{IRTest_summary}(x, ...)
}
\arguments{
\item{x}{An object returned from \code{\link{summary.IRTest}}.}

\item{...}{Additional arguments (currently non-functioning).}
}
\value{
Summarized texts on the console.
}
\description{
This function prints the summarized information.
}
\examples{
\donttest{
data <- DataGeneration(N=1000, nitem_P = 8)$data_P

M1 <- IRTest_Poly(data = data,
                  latent_dist = "2NM")

summary(M1)
}
}
\author{
Seewoo Li \email{cu@yonsei.ac.kr}
}
