% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_collapsing.R
\name{recategorize}
\alias{recategorize}
\title{Recategorization of data using a new categorization scheme}
\usage{
recategorize(data, new_cat)
}
\arguments{
\item{data}{An item response matrix}

\item{new_cat}{A list of a new categorization scheme}
}
\value{
Recategorized data
}
\description{
With a recategorization scheme as an input, this function implements recategorization for the input data.
}
