% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.irtest.R
\name{summary.irtest}
\alias{summary.irtest}
\title{Summary of the results}
\usage{
\method{summary}{irtest}(object, ...)
}
\arguments{
\item{object}{An \code{class == "irtest"} object obtained from either \code{\link{IRTest_Dich}}, \code{\link{IRTest_Poly}}, or \code{\link{IRTest_Mix}}.}

\item{...}{Other argument(s) passed on to summarize the results.}
}
\value{
A plot of estimated latent distribution.
}
\description{
These functions summarize the outputs (e.g., convergence of the estimation algorithm, parameter estimates, AIC, etc.).
}
